# examples/httponly/openapiv2/hellow/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "hellow/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "HellowService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/hellow": {
      "get": {
        "summary": "GetHellow handles the GET /hellow?name=Name endpoint",
        "operationId": "HellowService_GetHellow",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/webHellowResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "HellowService"
        ]
      },
      "delete": {
        "summary": "DeleteHellow handles the DELETE /hellow?name=Name endpoint",
        "operationId": "HellowService_DeleteHellow",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/webHellowResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "HellowService"
        ]
      },
      "post": {
        "summary": "Post a request body.",
        "operationId": "HellowService_PostHellow",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/webHellowResponse"
            }
          },
          "400": {
            "description": "Bad Request - Invalid Input.",
            "schema": {
              "$ref": "#/definitions/webHellowResponse"
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/webHellowRequest"
            }
          },
          {
            "name": "X-Request-ID",
            "description": "Unique request ID",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "HellowService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "webHellowRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "webHellowResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "data": {
          "type": "string"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "title": "HellowResponse represents the response for both POST and GET endpoints"
    }
  }
}